Organise Apply page test helpers
Similar to the Nunjucks template organisation in the previous commit, we organise our Apply page integration test helper classes to follow the same [section]/[task]/[page] as the classes they test
