+++ cypress_shared/pages/apply/accommodation-need/sentence-information/sentenceType.ts
@@ -1,6 +1,6 @@
 import type { TemporaryAccommodationApplication } from '@approved-premises/api'
 
-import ApplyPage from './applyPage'
+import ApplyPage from '../../applyPage'
 
 export default class SentenceTypePage extends ApplyPage {
   constructor(application: TemporaryAccommodationApplication) {
