+++ cypress_shared/pages/apply/index.ts
@@ -1,31 +1,31 @@
-import AccommodationRequiredFromDatePage from './accommodationRequiredFromDate'
-import AttachDocumentsPage from './attachDocumentsPage'
-import BackupContactPage from './backupContact'
-import CheckYourAnswersPage from './checkYourAnswersPage'
+import BackupContactPage from './accommodation-need/contact-details/backupContact'
+import PractitionerPduPage from './accommodation-need/contact-details/practitionerPdu'
+import ProbationPractitionerPage from './accommodation-need/contact-details/probationPractitioner'
+import AccommodationRequiredFromDatePage from './accommodation-need/eligibility/accommodationRequiredFromDate'
+import EligibilityReasonPage from './accommodation-need/eligibility/eligibilityReason'
+import ReleaseDatePage from './accommodation-need/eligibility/releaseDate'
+import SentenceTypePage from './accommodation-need/sentence-information/sentenceType'
+import CheckYourAnswersPage from './check-your-answers/check-your-answers/checkYourAnswersPage'
 import ConfirmDetailsPage from './confirmDetails'
-import CrsSubmittedPage from './crsSubmitted'
-import DtrDetailsPage from './dtrDetails'
-import DtrSubmittedPage from './dtrSubmitted'
-import EligibilityReasonPage from './eligibilityReason'
 import EnterCRNPage from './enterCrn'
-import FoodAllergiesPage from './foodAllergies'
 import ListPage from './list'
-import MoveOnPlanPage from './moveOnPlan'
-import OffenceDetailsPage from './offenceDetails'
-import OptionalOasysSectionsPage from './optionalOasysSections'
-import OtherAccommodationOptionsPage from './otherAccommodationOptions'
-import PractitionerPduPage from './practitionerPdu'
-import ProbationPractitionerPage from './probationPractitioner'
-import ReleaseDatePage from './releaseDate'
-import RiskManagementPlanPage from './riskManagementPlan'
-import RiskToSelfPage from './riskToSelf'
-import RoomSharingPage from './roomSharing'
-import RoshSummaryPage from './roshSummary'
+import MoveOnPlanPage from './placement-considerations/move-on-plan/moveOnPlan'
+import FoodAllergiesPage from './placement-considerations/requirements-for-placement/foodAllergies'
+import RoomSharingPage from './placement-considerations/requirements-for-placement/roomSharing'
+import CrsSubmittedPage from './referrals-and-documents/accommodation-referral-details/crsSubmitted'
+import DtrDetailsPage from './referrals-and-documents/accommodation-referral-details/dtrDetails'
+import DtrSubmittedPage from './referrals-and-documents/accommodation-referral-details/dtrSubmitted'
+import OtherAccommodationOptionsPage from './referrals-and-documents/accommodation-referral-details/otherAccommodationOptions'
+import AttachDocumentsPage from './referrals-and-documents/attach-documents/attachDocumentsPage'
+import OffenceDetailsPage from './risk-strength-and-need-information/oasys-import/offenceDetails'
+import OptionalOasysSectionsPage from './risk-strength-and-need-information/oasys-import/optionalOasysSections'
+import RiskManagementPlanPage from './risk-strength-and-need-information/oasys-import/riskManagementPlan'
+import RiskToSelfPage from './risk-strength-and-need-information/oasys-import/riskToSelf'
+import RoshSummaryPage from './risk-strength-and-need-information/oasys-import/roshSummary'
+import SupportingInformationPage from './risk-strength-and-need-information/oasys-import/supportingInformation'
 import SelectOffencePage from './selectOffence'
-import SentenceTypePage from './sentenceType'
 import StartPage from './startPage'
 import SubmissionConfirmation from './submissionConfirmation'
-import SupportingInformationPage from './supportingInformation'
 import TaskListPage from './taskListPage'
 
 export {
