+++ cypress_shared/pages/apply/placement-considerations/requirements-for-placement/roomSharing.ts
@@ -1,6 +1,6 @@
 import type { TemporaryAccommodationApplication } from '@approved-premises/api'
-import paths from '../../../server/paths/apply'
-import ApplyPage from './applyPage'
+import paths from '../../../../../server/paths/apply'
+import ApplyPage from '../../applyPage'
 
 export default class RoomSharingPage extends ApplyPage {
   constructor(application: TemporaryAccommodationApplication) {
