+++ cypress_shared/pages/apply/referrals-and-documents/accommodation-referral-details/crsSubmitted.ts
@@ -1,7 +1,7 @@
 import type { TemporaryAccommodationApplication } from '@approved-premises/api'
-import { hasSubmittedDtr } from '../../../server/form-pages/utils'
-import paths from '../../../server/paths/apply'
-import ApplyPage from './applyPage'
+import { hasSubmittedDtr } from '../../../../../server/form-pages/utils'
+import paths from '../../../../../server/paths/apply'
+import ApplyPage from '../../applyPage'
 
 export default class CrsSubmittedPage extends ApplyPage {
   constructor(application: TemporaryAccommodationApplication) {
