+++ cypress_shared/pages/apply/referrals-and-documents/accommodation-referral-details/otherAccommodationOptions.ts
@@ -1,6 +1,6 @@
 import type { TemporaryAccommodationApplication } from '@approved-premises/api'
-import paths from '../../../server/paths/apply'
-import ApplyPage from './applyPage'
+import paths from '../../../../../server/paths/apply'
+import ApplyPage from '../../applyPage'
 
 export default class OtherAccommodationOptionsPage extends ApplyPage {
   constructor(application: TemporaryAccommodationApplication) {
