+++ cypress_shared/pages/apply/risk-strength-and-need-information/oasys-import/riskToSelf.ts
@@ -1,7 +1,7 @@
 import { ArrayOfOASysRiskToSelfQuestions, TemporaryAccommodationApplication } from '@approved-premises/api'
-import paths from '../../../server/paths/apply'
+import paths from '../../../../../server/paths/apply'
 
-import ApplyPage from './applyPage'
+import ApplyPage from '../../applyPage'
 
 export default class RiskToSelf extends ApplyPage {
   constructor(
