+++ cypress_shared/pages/apply/risk-strength-and-need-information/oasys-import/roshSummary.ts
@@ -2,9 +2,9 @@ import {
   ArrayOfOASysRiskOfSeriousHarmSummaryQuestions,
   TemporaryAccommodationApplication,
 } from '@approved-premises/api'
-import paths from '../../../server/paths/apply'
+import paths from '../../../../../server/paths/apply'
 
-import ApplyPage from './applyPage'
+import ApplyPage from '../../applyPage'
 
 export default class RoshSummary extends ApplyPage {
   constructor(
