+++ cypress_shared/pages/apply/risk-strength-and-need-information/oasys-import/supportingInformation.ts
@@ -1,7 +1,7 @@
 import { ArrayOfOASysSupportingInformationQuestions, TemporaryAccommodationApplication } from '@approved-premises/api'
-import paths from '../../../server/paths/apply'
+import paths from '../../../../../server/paths/apply'
 
-import ApplyPage from './applyPage'
+import ApplyPage from '../../applyPage'
 
 export default class SupportingInformation extends ApplyPage {
   constructor(
